% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicproperties.R
\docType{package}
\name{basicproperties}
\alias{basicproperties}
\alias{basicproperties-package}
\title{Calculate a Basic Set of Time-Series Properties}
\description{
Calculate a Basic Set of Time-Series Properties
}
