% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate the mode of a vector}
\usage{
calc_mode(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{numeric} scalar denoting the most frequent value
}
\description{
Calculate the mode of a vector
}
\examples{
y <- arima.sim(model = list(ar = 0.8), n = 1000)
outs <- get_properties(y)

}
\author{
Trent Henderson
}
