% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantiles.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate all quantiles for an input vector}
\usage{
calculate_quantiles(y, quantiles = seq(0.01, 1, by = 0.01))
}
\arguments{
\item{y}{\code{numeric} vector of values}

\item{quantiles}{\code{numeric} vector of quantiles to calculate. Defaults to \code{seq(0.01, 1, by = 0.01)}}
}
\value{
\code{data.frame} of results
}
\description{
Calculate all quantiles for an input vector
}
\author{
Trent Henderson
}
