% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_coefficients.R
\name{fft_coefficients}
\alias{fft_coefficients}
\title{Calculate the first 100 FFT coefficients for a time-series vector as per the tsfresh package for Python}
\usage{
fft_coefficients(y, attrs = c("real", "imag", "abs", "angle"))
}
\arguments{
\item{y}{\code{numeric} vector of values}

\item{attrs}{\code{character} vector denoting which components to calculate. Can be one of \code{"real"}, \code{"imag"}, \code{"abs"}, or \code{"angle"}. Defaults to all through \code{c("real", "imag", "abs", "angle")}}
}
\value{
\code{data.frame} of results
}
\description{
Calculate the first 100 FFT coefficients for a time-series vector as per the tsfresh package for Python
}
\references{
Christ, M., Braun, N., Neuffer, J., and Kempa-Liehr A.W. (2018). Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh -- A Python package). Neurocomputing 307
}
\author{
Trent Henderson
}
