% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_properties.R
\name{get_properties}
\alias{get_properties}
\title{Calculate all features in the package on an input vector}
\usage{
get_properties(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{data.frame} that contains the summary statistics for each feature
}
\description{
Calculate all features in the package on an input vector
}
\examples{
y <- arima.sim(model = list(ar = 0.8), n = 1000)
outs <- get_properties(y)

}
\author{
Trent Henderson
}
