% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate tailedness of the value distribution}
\usage{
kurtosis(y)
}
\arguments{
\item{y}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate tailedness of the value distribution
}
\examples{
y <- stats::arima.sim(model = list(ar = 0.8), n = 1000)
kurtosis(y)

}
\author{
Trent Henderson
}
