% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_trend.R
\name{linear_trend}
\alias{linear_trend}
\title{Calculate the linear trend of a time series}
\usage{
linear_trend(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{numeric} scalar denoting the most frequent value
}
\description{
Calculate the linear trend of a time series
}
\examples{
y <- arima.sim(model = list(ar = 0.8), n = 1000)
outs <- linear_trend(y)

}
\author{
Trent Henderson
}
