% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_abs_above_3SD.R
\name{prop_abs_above_3SD}
\alias{prop_abs_above_3SD}
\title{Calculate the proportion of values in a time series greater than absolute 3*SD}
\usage{
prop_abs_above_3SD(y, na.rm = FALSE)
}
\arguments{
\item{y}{\code{numeric} vector of values}

\item{na.rm}{\code{Boolean} denoting whether missing values be removed. Defaults to \code{FALSE}}
}
\value{
\code{numeric} scalar denoting the proportion of values in \code{y} greater than absolute 3*SD
}
\description{
Calculate the proportion of values in a time series greater than absolute 3*SD
}
\examples{
y <- arima.sim(model = list(ar = 0.8), n = 1000)
outs <- prop_abs_above_3SD(y)

}
\author{
Trent Henderson
}
