% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skewness}
\alias{skewness}
\title{Calculate adjusted Fisher-Pearson coefficient of skewness}
\usage{
skewness(y)
}
\arguments{
\item{y}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate adjusted Fisher-Pearson coefficient of skewness
}
\examples{
y <- stats::arima.sim(model = list(ar = 0.8), n = 1000)
skewness(y)

}
\author{
Trent Henderson
}
