% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_centroid.R
\name{spectral_centroid}
\alias{spectral_centroid}
\title{Calculate the spectral centroid of a time series}
\usage{
spectral_centroid(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{numeric} scalar denoting the spectral centroid obtained through the FFT
}
\description{
Calculate the spectral centroid of a time series
}
\examples{
y <- arima.sim(model = list(ar = 0.8), n = 1000)
outs <- spectral_centroid(y)

}
\author{
Trent Henderson
}
